enable CMAKE_BUILD_TYPE MinSizeRel
The CMAKE_BUILD_TYPE MinSizeRel is currently not recognized so that
the build system falls back to Release. In contrast to Release,
MinSizeRel advises the compiler to optimize for size instead of speed
which is particularly interesting on micro controllers. For GCC
CMake will typically specify the option -Os instead of -O3.