Bypass Linking for CMake Compiler Test
Some generators throw in a `try_compile()` sanity check that will fail
trying to link and run an exe on the host platform since this is a
cross-compiler.  This behavior was seen on Windows.

Since linking and running won't work with this toolchain on the host,
change the CMAKE_TRY_COMPILE_TARGET_TYPE to "STATIC_LIBRARY".